@@ -1,15 +1,21 @@
 module github.com/junegunn/fzf
 
 require (
-	github.com/gdamore/tcell v1.4.0
-	github.com/mattn/go-isatty v0.0.12
-	github.com/mattn/go-runewidth v0.0.9
-	github.com/mattn/go-shellwords v1.0.10
-	github.com/saracen/walker v0.1.1
-	golang.org/x/crypto v0.0.0-20201016220609-9e8e0b390897
-	golang.org/x/sync v0.0.0-20201020160332-67f06af15bc9 // indirect
-	golang.org/x/sys v0.0.0-20201026173827-119d4633e4d1
-	golang.org/x/text v0.3.3 // indirect
+	github.com/gdamore/tcell/v2 v2.5.4
+	github.com/mattn/go-isatty v0.0.17
+	github.com/mattn/go-runewidth v0.0.14
+	github.com/mattn/go-shellwords v1.0.12
+	github.com/rivo/uniseg v0.4.2
+	github.com/saracen/walker v0.1.3
+	golang.org/x/sys v0.0.0-20220811171246-fbc7d0a398ab
+	golang.org/x/term v0.0.0-20210927222741-03fcf44c2211
 )
 
-go 1.13
+require (
+	github.com/gdamore/encoding v1.0.0 // indirect
+	github.com/lucasb-eyer/go-colorful v1.2.0 // indirect
+	golang.org/x/sync v0.0.0-20220722155255-886fb9371eb4 // indirect
+	golang.org/x/text v0.5.0 // indirect
+)
+
+go 1.17
